+++ github_repo_changes.rb
@@ -6,11 +6,10 @@ class GithubRepoChanges
   attr_reader :version_regex
   def initialize(config_file: 'config.yml',
                  repo: nil,
-                 tags_filter: 'onlyoffice-documentserver-enterprise-',
                  version_regex: /\d*\.\d*.\d*-\d*/)
     @config = YAML.load_file(config_file)
     @repo = repo
-    @tags_filter = tags_filter
+    @tags_filter = @config['tags_filter']
     @version_regex = version_regex
     Octokit.configure do |c|
       c.login = @config['github_user']