+++ lib/telegra_github_changes_bot.rb
@@ -1,0 +1,2 @@
+require 'telegram/bot'
+require_relative 'telegram_github_changes_bot/github_repo_changes'