+++ bot.rb
@@ -7,18 +7,25 @@ config['repos'].each do |cur_repo|
 end
 
 Telegram::Bot::Client.run(config['telegram_bot_token']) do |bot|
-  bot.listen do |message|
-    case message.text
-    when %r{/get_changes.*}
-      text = ''
-      repos_changes_array.each do |cur_repo|
-        version = message.text[cur_repo.version_regex]
-        text += cur_repo.link_to_changes(version)
+  begin
+    bot.listen do |message|
+      case message.text
+      when %r{/get_changes.*}
+        text = ''
+        repos_changes_array.each do |cur_repo|
+          version = message.text[cur_repo.version_regex]
+          text += cur_repo.link_to_changes(version)
+        end
+        text = 'There is no changes for latest version' if text.empty?
+        bot.api.send_message(chat_id: message.chat.id,
+                             text: text,
+                             parse_mode: 'Markdown')
       end
-      text = 'There is no changes for latest version' if text.empty?
-      bot.api.send_message(chat_id: message.chat.id,
-                           text: text,
-                           parse_mode: 'Markdown')
     end
+  rescue Telegram::Bot::Exceptions::ResponseError => e
+    if e.error_code.to_s == '502'
+      puts 'Telegram raised 502 error. Pretty normal, ignore that'
+    end
+    retry
   end
 end