+++ lib/telegram_github_changes_bot/github_repo_changes.rb
@@ -1,40 +1,37 @@
 require 'octokit'
 require 'yaml'
+require_relative 'github_repo_changes/github_repo_changes_helper'
 
 # Class for describing changes for github repo
 class GithubRepoChanges
+  include GithubRepoChangesHelper
   attr_reader :version_regex
+  attr_accessor :old_tag
+  attr_accessor :new_tag
   def initialize(config_file: 'config.yml',
                  repo: nil)
+    return if config_file.nil?
     @config = YAML.load_file(config_file)
     @repo = repo
-    @tags_filter = @config['tags_filter']
     @version_regex = /#{@config['version_regexp']}/
     Octokit.configure do |c|
       c.login = @config['github_user']
       c.password = @config['github_user_password']
     end
-    Octokit.auto_paginate = true
-  end
-
-  # @param string [String] string for check
-  # @return [True, False] check if tag matches regexp
-  def tag_match?(string)
-    (string =~ /#{@tags_filter}\d/).nil?
   end
 
   def tag_names
     tags = []
     Octokit.tags(@repo).each do |current_tag|
-      tags << current_tag['name'] unless tag_match?(current_tag['name'])
+      tags << current_tag['name']
     end
     tags
   end
 
-  def fetch_latest_tags
+  def fetch_tags
     tags = tag_names
-    @new_tag = tags[0]
-    @old_tag = tags[1]
+    @new_tag = tags[0] unless @new_tag
+    @old_tag = tags[1] unless @old_tag
   end
 
   def changes_url
@@ -46,11 +43,10 @@ class GithubRepoChanges
     changes[:files].empty?
   end
 
-  def link_to_changes(specific_version = nil)
-    fetch_latest_tags
-    @old_tag = "#{@tags_filter}#{specific_version}" unless specific_version.nil?
+  def link_to_changes
+    fetch_tags
     return '' if changes_empty?
-    "[#{@repo} changes v.#{@old_tag.gsub(@tags_filter, '')} -"\
-    "v.#{@new_tag.gsub(@tags_filter, '')}](#{changes_url})\n"
+    "[#{@repo} changes #{@old_tag}..."\
+    "#{@new_tag}](#{changes_url})\n"
   end
 end