+++ spec/spec_helper.rb
@@ -1,0 +1,1 @@
+require 'telegram_github_changes_bot'