+++ lib/telegram_github_changes_bot/github_repo_changes.rb
@@ -1,52 +1,51 @@
 require 'octokit'
 require 'yaml'
 require_relative 'github_repo_changes/github_repo_changes_helper'
+require_relative 'github_repo_changes/ref_helper'
 
 # Class for describing changes for github repo
 class GithubRepoChanges
   include GithubRepoChangesHelper
+  include RefHelper
   attr_reader :version_regex
-  attr_accessor :old_tag
-  attr_accessor :new_tag
+  attr_accessor :old_ref
+  attr_accessor :new_ref
   def initialize(config_file: 'config.yml',
                  repo: nil)
-    return if config_file.nil?
-    @config = YAML.load_file(config_file)
+    init_github_access(config_file)
     @repo = repo
-    @version_regex = /#{@config['version_regexp']}/
     Octokit.configure do |c|
-      c.login = @config['github_user']
-      c.password = @config['github_user_password']
+      c.login = @user_name
+      c.password = @user_password
     end
-  end
-
-  def tag_names
-    tags = []
-    Octokit.tags(@repo).each do |current_tag|
-      tags << current_tag['name']
-    end
-    tags
-  end
-
-  def fetch_tags
-    tags = tag_names
-    @new_tag = tags[0] unless @new_tag
-    @old_tag = tags[1] unless @old_tag
   end
 
   def changes_url
-    "https://github.com/#{@repo}/compare/#{@old_tag}...#{@new_tag}"
+    "https://github.com/#{@repo}/compare/#{@old_ref}...#{@new_ref}"
   end
 
   def changes_empty?
-    changes = Octokit.compare(@repo, @old_tag, @new_tag)
+    changes = Octokit.compare(@repo, @old_ref, @new_ref)
     changes[:files].empty?
   end
 
   def link_to_changes
-    fetch_tags
+    fetch_refs
+    return "There is no #{@old_ref} in #{@repo}\n" unless ref_exist?(@old_ref)
+    return "There is no #{@new_ref} in #{@repo}\n" unless ref_exist?(@new_ref)
     return '' if changes_empty?
-    "[#{@repo} changes #{@old_tag}..."\
-    "#{@new_tag}](#{changes_url})\n"
+    "[#{@repo} changes #{@old_ref}..."\
+    "#{@new_ref}](#{changes_url})\n"
+  end
+
+  private
+
+  def init_github_access(config)
+    @user_name = ENV['GITHUB_USER_NAME']
+    @user_password = ENV['GITHUB_USER_PASSWORD']
+    return unless File.exist?(config)
+    @config = YAML.load_file(config)
+    @user_name = @config['github_user']
+    @user_password = @config['github_user_password']
   end
 end