+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased (master)
 ### New Features
 * Show version number in changes link
+* Ability to get changes from specific version: ask `/get_changes 3.8.0-208`
 
 ## 0.0.2
 ### New Features