+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### New Features
 * Show version number in changes link
 * Ability to get changes from specific version: ask `/get_changes 3.8.0-208`
+* Move setting tag filter to config file
 
 ## 0.0.2
 ### New Features