+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Show version number in changes link
 * Ability to get changes from specific version: ask `/get_changes 3.8.0-208`
 * Move setting tag filter to config file
+* More string checking for filter tags, after tags filter should be version numbers
 
 ## 0.0.2
 ### New Features