+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Move setting tag filter to config file
 * More string checking for filter tags, after tags filter should be version numbers
 * Move regexp for version to config file
+* Some refactor
 
 ## 0.0.2
 ### New Features