+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change log
 
 ## Unreleased (master)
+
 ### New Features
+
 * Show version number in changes link
 * Ability to get changes from specific version: ask `/get_changes 3.8.0-208`
 * Move setting tag filter to config file
@@ -15,8 +17,11 @@
 * Force pagination - without it some data meay be lost
 
 ## 0.0.2
+
 ### New Features
+
 * Show only repos with new changes
 
 ## 0.0.1
+
 * Initial Release