+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Use `refs` instead of `tags`
 * Show more detailed message if specified ref - is the latest
 * Force pagination - without it some data meay be lost
+* Do not install test gems in dockerfile
 
 ## 0.0.2